Move createBlock under BlockGenerator trait under tests
